backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-dotenv==1.0.0
openai==1.3.0
firebase-admin==6.2.0
pytest==7.4.3
httpx==0.25.1
python-multipart==0.0.6
pydantic==2.4.2
requests==2.31.0
google-generativeai==0.3.1 